Fix adaptive_grain function signature

+++ stgfunc/noise.py
@@ -27,7 +27,7 @@ GrainerFuncGenerator = Callable[[float, float, int, bool], GrainerFunc]
 @disallow_variable_resolution
 def adaptive_grain(
     clip: vs.VideoNode, strength: float | List[float] = 0.25, size: float = 1, sharp: int = 50, static: bool = False,
-    luma_scaling: int = 12, grainer: Grainer | GrainerFuncGenerator | str | None = Grainer.AddGrain,
+    luma_scaling: float = 12, grainer: Grainer | GrainerFuncGenerator | str | None = Grainer.AddGrain,
     fade_edges: bool = True, tv_range: bool = True, lo: int | None = None, hi: int | None = None,
     protect_neutral: bool = True, seed: int = -1, show_mask: bool = False, temporal_average: int = 0, **kwargs: Any
 ) -> vs.VideoNode:
